Refactor configuration setup
The principle of this refactoring is to simplify the configuration of
the application. A canonical configuration file location was
established at /etc/arbeitszeitapp/arbeitszeitapp.py.  The
ARBEITSZEITAPP_CONFIGURATION_PATH allows to specify an alternate
configuration file location.

The file arbeitszeit_flask/configuration_base.py now is a location for
default settings. Instead of importing the file from other settings
files it is automatically loaded on application startup. All
configuration options from this file can be overridden by the user
specified configuration.

The documentation on app config was extended.