Change DB interface for creating plans
Previously the DB interface for creating a plan accepted simply a PlanDraft.id
to create a plan from the PlanDraft that the id referenced. With this change
there is a dedicated methods for creating a plan independently from PlanDraft.
The goal is to have a more flexible DB interface which is achieved by this
change since now Plan records can be created without the need for a preexisting
PlanDraft. Secondly the removes implicit logic from the DB. The previous
approach relied on the DB implementation filling in the correct fields in the
Plan from the PlanDraft that the Plan was created from. Now the business logic
has complete control over what fields Plan records are created with. This means
that we can test the usage of correct values once in the business logic instead
of having to test this in the DB implementation.