Update development dependencies
We updated the nixpkgs version used for the development setup.  This
updated the version of black and mypy that are used in
development. That's why we needed to change some formatting and some
type annotations to make the tests pass. Due to
https://github.com/NixOS/nixpkgs/pull/211205 the development
dependency set shrunk by a couple of packages which should speed up
development ever so slightly.