Changed labour certificate payout modalities
With this change the labour certificate payout is changed such that
certificates going to the A account of companies get transferred at
the end of every full day the plan is active. Before this change the
certificates got transferred at the start of each day a plan was
active.

Also a bug was fixed were a company would get too many certificates if
the payout use case was triggered exactly on the same time as the plan
would expire.

Thirdly we introduced a method to filter plans from the DB such that
only plans with pending payouts for labour certificates are selected.