Change tests.use_cases.test_get_company_summary to use use case interactions
This changes the use case tests for GetCompanySummary. The test data
generation was previously done via "manually" created
transactions. Now we use use case interactions only to facilitate the
test data setup. This has the advantage of being more rebust in the
face of changes to implementation details in the business logic.