Remove redundant `account_type` model from database
This change removes the redundant field "account_type" from the
database. This field used to contain an identifier for the accounts
purpose, e.g. "p" for fixed means account or "member" for consumer
account. This information is redundant since the purpose of an account
can already be determined without that identifier by looking at how
the account is linked to a company, member or social accounting.