Fix bug in GetCompanySummary use case, fixes #671
This patch changes the DB interface slightly. Previously there was a
method `get_sales_balance_of_plan` part of the `TransactionRepository`
interface. This method was intended to look at all the transactions
for a certain product (denoted by its plan id) and yield its overall
contribution to the PRD account. This was bugged in the production
implementation (see
https://github.com/arbeitszeit/arbeitszeitapp/issues/671 for detail on
that). Instead of simply fixing that function I decided to remove it
entirely. The reason for this is that this method encoded business
logic unnecessarily into the database interfacing code. The interface
`arbeitszeit.repositories.TransactionResult` was extended by the
method `that_were_a_sale_for_plan` to replace the now deleted
`get_sales_balance_of_plan` method. The implementation of this method
`arbeitszeit_flask.database.repositories.TransactionQueryResult.that_were_a_sale_for_plan`
could be optimized further but I think that it is good enough for now.