Refactor arbeitszeit.repositories.AccountantRepository
We removed the `has_accountant_with_email` and `get_by_id` methods
from AccountantRepository and created a proper AccountantResult
interface that offers replacements for the removed methods. Also the
`get_all_accountants` method was renamed to `get_accountants` to be
more consitent with the naming scheme of the database gateway.