Refactor DB interface regarding Cooperation objects
This refactoring changes the DB interface for retrieving Cooperation
objects/rows from the DB. The previous CooperationRepository was
replaced by the approriate methods in the DatabaseGateway interface
and implementations for the new CooperationResult interface. Tests for
the production implementation of CooperationResult were
implemented.

With this change the Cooperation entity does not store the coordinator
company as a field anymore but instead only has the coordinator UUID
as a reference. This was done to improve DB performance since it
avoids querying the coordinator company when it is not needed. The
CooperationResult interface provided a method to yield coordinator
companies for cooperations on demand as a replacement.