Rework DB interface for handling CompanyWorkInvites
This change is purely a refactoring. It reworks the DB interface for
handling CompanyWorkInvite records. The reworked interface is more in
line with the "new style" interface by providing a query result that
can be manipulated without issuing any queries to the underlying
database.

Secondly the CompanyWorkInvite entity now only holds UUIDs referencing
the issuing company and the addressed member account.