Implement override protection for company passwords
Before this change it was possible to change a companies password by
registering a member account with the same email address. This
possible attack vector is eliminated with this change. The use will
see a minimal error message if they try to register a company with the
email address of a preexisting member account and the provided
password does not match the members password.