Encode redirection url after login in cookie instead of url query
When accessing a page that requires authentication we redirect users
back to the start page and notify them that they need to log
in. Previously we would encode the original target url in a url
query. With this change we encode the original target location in the
session cookie instead.

Also the code for authenticating client requests was moved from the
flask specific folder `arbeitszeit_flask` to the folder for framework
agnostic web code `arbeitszeit_web`. This allows us to test the
redirection logic more easily in later changes to our code.