Implement AccountOwnerRepository as method of AccountResult
We removed the AccountOwnerRepository interface and its
implementations completely and integrated the provided functionality
into the AccountResult interface.  This change should not increase or
decrease the amount of DB calls made in the use cases.

Secondly the SocialAccounting entity was changed to contain only a
UUID reference to its associated account instead of the account
entity. This change decreases the amount of DB calls when the user
views their transaction overview in the UI.