Move token generation logic from business layer to web layer
This change moves all the web token generation and validation from the
business layer (`/arbeitszeit`) to the web layer
(`/arbeitszeit_web`). This change makes it much easier to test the
business logic since it is not concerned with input validation. The
input validation logic itself is also easier to test because of this
change.

As part of this change two new controllers were introduced:
ConfirmMemberController and ConfirmCompanyController.  These new
controllers are concerned with extracting the email address encoded in
the token and validating that the token is not expired yet.

For better transparency all presenter interfaces declared in the
business layer (`/arbeitszeit`) were moved to
`/arbeitszeit/presenters.py`.