Database: Implement method to join Transactions with their sender and receiver
This changeset introduces a new method on the TransactionResult
interface: joined_with_sender_and_receiver.  This method was
introduced to optimize the GetCompanyTransactions use case.  Before
this change there were two separate DB query for each transaction made
from or to a company.  With the new method there is only one single DB
query made for all the transactions from/to a company combined. This
speeds up the execution time for the use case by orders of magnitude.