tests: MemberGenerator.create_member does not accept confirmed_on
MemberGenerator.create_member was changed to not accept the
confirmed_on argument anymore. This change was done since this feature
cannot be supported with a monotonic flow of time (meaning that the
hands of the clock only move forwards). Users that require to set a
specific confirmed_on timestamp in test cases should instead use the
`FakeDatetimeService` to manipulate the clock
(`freeze_time(timestamp)`, `advance_time(timedelta)`).