database: Move email address specific info into new email table
This change creates a new database table "email". The purpose of this
table is to store information associated with email addresses. This
was done in preparation to implement a "change email address" feature
for users.

This table will allow us to separate the steps of confirming an email
address and changing a users email address.