tests: Rework tests for ShowPRDAccountDetailsUseCase
This change reworks the tests for the
ShowPRDAccountDetailsUseCase. The basic idea of this change was to
remove all explicit references to the transaction generator and the
transaction repository. Instead the creation of the test fixtures is
done by invoking the proper use cases for for "buying" products from a
company. This should make the tests more robust in the face of changes
to the DB structure.