Implement get_member_account_details view
With this change a new view is introduced. This view allows users
logged in as a member to check their basic account data. For now this
is limited to their user id and their email address. As it stands
right now this view is pretty useless since the same information is
also present on the member dashboard view. The newly introduced view
can serve as a basis for implementing new functionality to change a
users password and email address.