arbeitszeit: Change labour certificate handling
This commit changes that handling of labour certificates
significantly. Previously labour certificates were granted to
companies which would distribute those certificates among
themselfs (their workers).

After this change the companies wont receive any certificates but
instead get a labour time credit to their labour account. The amount
of this credit is identical to what companies planned. Instead of
distributing certificates the companies now register hours worked with
public accounting. The app will then automatically subtract amount of
hours worked from the labour account of the company and grant
certificates directly to the worker how performed the labour.

Wording in the frontend and renaming of identifiers in the code will
be done in a future commit.