Refactor arbeitszeit.transaction and get_member_account use case
This refactoring removes several obsolete methods of the
UserAccountingService and the GetUserAccount use case. More
importantly the logic for determining the "transaction type" in
arbeitszeit.accounting was reworked into a rule engine. This rule
engine allows easier review and modification of logic that governs
transaction types. The rules are layed out in a tabulated form so that
readers see easily how to add new rules or edit existing ones.

The author of this change took care to implement the same behavior as
there was before. That's why the rules "look" a little bit wrong,
especially for clauses where we match any receiving account.