Make QueryCompaniesRequest a concrete data structure
This change changes the QueryCompaniesRequest from an abstract base
class into a concrete data structure.  The old approach was just not
working out as intended. The new approach removes the need for any
subclassing or interface implementations. Callers simply construct an
instance of the "dataclass" QueryCompaniesRequest if they want to
engage the query_companies use case.