Optimize DB access in query_plans use case
This commit optimizes the database access in the query_plans use
case. The use case will now fetch all the required data to calculate
the plan listing in one single query to the database. This was
achieved by implementing a new method on the database gateway that
will join plans with their planning company and the planning
information for all plans that are cooperating with the plan in
question (which might be 0).

The changes in this commit improved the benchmark timing on the
authors machine for the query_plans_sorted_by_activation_date
benchmark from ~5 seconds down to ~1.5 seconds.