arbeitszeit_benchmark: Commit to DB and flush session on start up
With this change all currently existing benchmark tests will commit
any pending changes from the benchmark setup to the DB and flush the
database session before running the benchmarks. This is done to ensure
that behavior similar to the production environment. Without
committing and flushing we might actually not test the database but
rather the sqlalchemy session cache, which is pointless.