Introduce CoordinationTenures (2): database-level
Introduces the database model `CoordinationTenure`.  It replaces the column `coordinator` of the table `Cooperation`, but is more powerful, because it keeps track of start_date of the coordination tenure. One cooperation can now have a history of CoordinationTenures. This is important e.g. for allowing to pass coordination to other companies.