Fix bug in routes /member/unconfirmed and /company/unconfirmed
This commit addresses a bug in two routes from the `auth` blueprint,
namely `/member/unconfirmed` and `/company/unconfirmed`.  The root
cause for the issue (apart from not having proper tests) is that
previously we would check if a member/company is confirmed and then
redirect them to the start page if that's the case. Now the logic is
that we check explicitly if the current user is an unconfirmed
member/company and redirect them otherwise.

Also tests were written to check for this behavior.