Refactor tests.use_cases.repositories
The refactoring performed in this change makes it easier to deal with
the different indices used in the mock repository for the use case
tests. This was done in 3 steps:

Define an `Index` class. This new `Index` class makes it clear that an
index is mapping from a "hashable" to a set of "hashables". Values can
be added, removed and gotten from the index.

Combine all indices into one `Indexes` object. Instances of
`EntityStorage` now have an attribute `indices` that holds all
available indices. This should make it easier for the
reader/programmer to discover what indices already exist or where to
add a new index if necessary.

Adjust all the queries to the new system.