Implement type stub generation when updating dependencies
This change introduces a new step in our dependency update program:
Automatic type stub generation for development dependencies. The
benefit of this is that the type checker as more information avaiable
when analysing our code. This is especially useful when checking
arity (e.g. the amount of arguments/parameters a function has) of
called functions/methods from dependencies.