PayConsumerProduct -> RegisterPrivateConsumption
Rename the use case `PayConsumerProduct` to `RegisterPrivateConsumption`. This is the accurate description of what is happening in this use case. We were using the verb "pay" in order to be more user friendly, but we noticed that it is confusing. Also rename controller, presenter, view, route, form, url_index, tests, readme, user manual.