Remove login info from Member, Company, Accountant records
This change changes the API and data records for `Member`, `Company`
and `Accountant` user objects.  Those 3 classes no longer have login
information associated with them.  There is a new record type
`AccountCredentials` which holds information on password hashes and
email addresses. The database interface and implementations were
changed to accommodate this. For a birds eye view the behavior of the
application should not have changed.