Implement RequestEmailAddressChange use case
The use case RequestEmailAddressChange implemented with this change
should allow users to request changing their email address. The change
request is only granted if the new email address is not already taken
by another account in the system. The use case does deliberately not
check for authorization since it is very likely that we want to enable
admins to do this for other users in the future.  On success the use
case will send a notification that the email address change is
requested. The actual change must be implemented in a separate use
case.