Implement RequestEmailAddressChangeController for the web
The implemented controller is intended for taking a web request from a
user and constructing the appropriate use case request for the
RequestEmailAddressChangeUseCase from it. It does so by taking the
user id and role and retrieving their email address from the
database. The new email address is parsed from a form field passed
into the controller.