Disable mypy error about using abstract types
Before this change it was not really possible to use our dependency
injection system in conjuntion with mypys type checking. The main
reason for this was that mypy would complain if we specified abstract
types as concrete values. Here is an example:

```python
class Interface(Protocol):
    def method(self) -> None:
        ...

instance = injector.get(Interface)  # mypy considers this illegal
```

This change disables these kinds of errors.  The main benefit of this
is that we can delete lots of comments that would disable mypy because
of this error.