Simplify type signatures on DatabaseGateway et al
Before this change we made use of a crutch when it comes to type
signatures. This change uses the new `Self` typing feature of python
in all `*Result` types for database interfacing.  The type signatures
should be easier to understand and reproduce with this change. This
simplification does come with a cost though.  After this change we
generally disallow that a specific subtype of `QueryResult` may yield
another subtype of QueryResult that is not a subtype of itself. The
author considers this trade off worthwhile since he believes that
something in the design process must have gone horribly wrong if the
need for such a monstrosity ever arises.