Implement basic presenter for RequestEmailAddressChange use case
The presenter implemented here does two things:

1) It generates a warning that is displayed to the user when the
   request has failed.
2) It figures out the appropriate redirect when the request was
   granted.

However it is not integrated with the web frontend yet. This will be
done in a future change.

Further the base test class `tests.www.base_test_case.BaseTestCase`
now has three additional attributes `notifier`, `translator` and
`url_index` which can be accessed by all presenter and controller
tests should those need to do so. The rational behind this change is
that it is kind of noisy and tedious establish those objects in every
second test. This made it necessary to adjust many tests to not
override those new read-only attributes.