Rework email sending interface into single `Protocol`
This change is intended to preserve the observable behavior on all
levels of the application. It is more of a structural change. The goal
of this change is to make it easier to discover what kinds of emails
can be sent by the application and how to add new kinds of email to
the code.

Before this change all of the interfaces that the business logic would
need to send emails were created ad hoc and not really well organized
in the code. Also the class names of these interfaces were pretty long
and unwieldy. Another complaint with the code was that it was not at
all straight forward to test code that would rely on these
interfaces. Usually one would need to create a mock implementation for
each of those interfaces.

Now there is only one single interface involved in sending
email. Different kinds of emails are now differentiated by dedicated
`dataclass` types, which are data collections/degenerate classes
without methods. The advantage of this change is that there can be one
single mock implementation that handles all kinds of email message
kinds. Also it is much clearer now how to implement a new kind of
email message since there are proper type definitions that *should* be
self documenting.

The adjustments that were made to the production code and the tests
are pretty minor with two notable exceptions: The test_url_index.py
and test_signup_accountant_view.py tests which involve creating
invitation tokens for accountants. Instead of trying to come up with a
"proper" solution of obtaining a registration token we simply generate
one on the fly via the arbeitszeit_flask.token.TokenService. This was
deemed to be okay since the tests in question probe the flask
implementation which also hosts the TokenService that was used.