Refactor email sending logic when company requests cooperation
This change entirely concerns the RequestCooperation use case. The
goal of this change was to preserve the current behavior of the
application and instead make the source code more managable/legile.

Before this change the code was set up such that the
RequestCooperationPresenter was responsible for sending a notification
email to the cooperations coordinator while generating the http
respons for the requester. This is confusing/unexpected since all
other email notifications are sent out explicitly via a trigger in the
appropriate use case object.

This change moves the trigger for sending an email to the coordinator
of a cooperation part of the use case object.

As part of this change some minor refactoring was also done to the
test logic. The FakeEmailSender was renamed to FakeEmailService since
it is an implementation of the EmailService interface and not the
EmailSender interface.