Add proper greeting to accountant notifications
This does (unfortunately) two separate things:

Add a proper greeting to the email that accountants receive when a
plan is filed.

Refactor accountant notification system slightly such that the
accountant email address is retrieved inside of the
FilePlanWithAccountingUseCase instead of the presenter. This makes the
`UserAddressBook` obsolete. Consequently this interface and its
implementations and tests were removed.