Implement EmailChangeConfirmation presenter
This change implements a rudimentary presenter for email change
notifications. This functionality is not yet integrated into the
application since multiple pieces of this remain to be implemented.

The test implementation is not ideal since we interrogate the email
content via its expected implementation. Therefore the tests for email
content are expected to be quite fragile/break often.