Remove need for `setup.py`
This change removes the need for a `setup.py` file in our
repository. To facilitate this we needed to move away from using the
setuptools navigation of the `Babel` package. To that end we
implemented a small maintenance script that totally replaces the setup
tools frontend (python setup.py compile_messages, ...). While we were
at it we fixed a bug in our distribution logic which prevented source
and binary distributions to include the compiled translation
files (.mo files).

Further we now include the translations template file `messages.pot`
in the source code. It was a mistake to not track it since it
basically serves as the configuration for all the .po files that we
derive from it. Without it we would need to re-establish all metadata
about the translations everytime we create new versions of the .po
files, e.g. the project name, the authors email, etc.