Improve logging output for update_dependencies program
Previously the logging output of the mainenance program
`arbeitszeit_development.update_dependencies` was fixed. The user was
not able to control the amount of logging when running this
program. With this change we introduced a `-v` flag to the program by
which the user can increase the logging output. Currently the flag
will change the output such that all subcommands being run are printed
to stderr which should help to inform the user what is going on. This
will be useful for debugging and understanding the program.

The previous design of this program made it hard to control the
logging output globally since most of the subroutines involved in this
program were run as separate subprocesses. This made it hard and
inconvenient to globally control the logging in those subroutines. The
design was changed such that all subroutines run as part of the main
process. To archieve that we had to enable some subroutines to run
the required command lines tools within the nix development shell. For
this purpose we introduced the `SubprocessRunner` protocol.

The `SubprocessRunner` protocol allows clients to run arbitrary
subprocesses. The runner objects themselves are responsible for
constructing the approriate environment. This allows us to separate
the commands that need to be run from the environment they are run in.