Introduce RequestCoordinationTransferUseCase
This commit introduces a new use case, RequestCoordinationTransferUseCase (see issue #317).  It allows current coordinators of a cooperation to request the transferal of the coordination role to another company. A mail will be sent out. Together with this use case comes a new `record`, called CoordinationTransferRequest.  Note that this use case fails if the requesting company (i.e. coordination tenure) already has requested a transfer to another company. This behaviour makes probably sense, because we expect companies to agree "in real life" which company will take over an existing coordination role - there will be no needfor a bunch of requests to be sent out. Note also that the newly created CoordinationTenureGenerator creates the coordination tenure via direct access to the database_gateway, because we have not yet all necessary use cases in place to do this via use cases.
The missing counterpart of this use case, called probably AcceptCoordinationTransferUseCase, is still to be written. Also the www implementation is still missing. [Labour time: 6h]