Add attr "end_time" to list of coordinators
The ListCoordinationsOfCooperationUseCase returned already the start_time of the coordination tenure, but not the end_time. The code in this commit calculates the end_time: The latest tenure in a cooperation has no end_time, the older coordination tenures have as their end_time the start_time of the more recent coordination tenure.