Web and db code for listing coordinations of coops
This commit adds the necessary web and db  code for listing the coordinators of a cooperation. The use case existed already.
Note that listing the coordinators is currently not a feature for users logged in as accountants (they are not able to access cooperation summary neither, see issue #853).

After this change, on the cooperation summary page there will not only be displayed the current coordinator, but also a link to the "history of coordinators". On this "history" page there are listed all coordinators of the coop, together with start and end dates.