Change behavior of "renew plan" button
Before this change the "renew plan" button would load the "create
draft" form and only after the user would click the "save draft"
button a new draft would be saved.

After this change a new draft will immediately be created when user
clicks "renew plan" button. The user will then be redirected to the
draft details view for that newly created draft.

This change allows us to extend the "renew plan" feature easily to
other plans then just expired ones from the planning company. It would
be very easy to allow the user to "clone" plans from other companies
or create drafts from currently active plans via the means provided by
this change. Also the individual parts of this implementation can be
tested more easily and rigorously as there is a clear separation
between the draft creation and the draft editing features of the
application. The coupling between those two systems is minimal.