Move user account details view into newly created user blueprint
Before this change we had to duplicate some of the web code for views
that should be available to members, accountants and companies
alike. This was because there was no mechanism in place to define
those views as available for all user types.

After this change there is a new flask blueprint "user" that allows
the definition of views that are shared between the different user
roles. The views `get_member_account_details`,
`get_company_account_details` and `get_accountant_account_details`
were moved to this new blueprint as they were an example of the above
mentioned duplication.