Have only member blueprint use `/member/` url prefix
Before this change there were a couple of routes in `auth` blueprint
that were using `/member/` in there urls as a prefix. This commit
changes those routes to not use the prefix in question anymore. This
allows us to use `/member` as a fixed url prefix for the `member`
blueprint.