Add @commit_change decorator to one company route
This commit adds the `commit_changes` decorator to the `main_company.request_coordination_transfer`route. Before this commit, the integration tests did succeed even though in production environment the changes of the POST request would not be persisted in the database. I added more accurate http return codes and improved integration testing. Still I was not able to cover this bug (db objects added to the db session but not persisted) in our flask integration test code.