Fix responsiveness problem on start page
Currently our start page is not responsive: Elements overflow the viewport horizontaly. It has to do with the duplication of the "hero-body" element in `start.html`. In general, our jinja inheritance was unnecessarily complicated, now it is more simple. I deleted the jinja "content_class" from the base template, because it is currently not needed. Also the  "body" and the "navigation" blocks were not needed, as the corresponding contents always appear in all our templates on top of the html `<body>` section.