Implement basic framework for parsing www form fields
Before this change there was no guidance nor a pattern one could
follow to implement www form field parsing. With this change a basic
framework for form field parsing is introduced.

The new functionality is located under `arbeitszeit_web.fields`. There
is now a function `parse_formfield` that takes a `FormField` and a
`FieldParser[T]` to produce an `Optional[T]` while attaching potential
error messages to the form field. A `FieldParser` can be any
`Callable` that receives a single string argument and produces
`ParsingSuccess[T] | ParsingFailure`.

The newly introduced framework was used to rework some www controllers
which already performed form field parsing.

One of the designs that were considered was an exception based control
flow a la django with its
[ValidationError](https://docs.djangoproject.com/en/4.2/ref/forms/fields/#django.forms.Field.clean).
This approach was rejected as it used exceptions as a control flow
mechanism which is only slightly less footgun-y than GOTO. Exceptions
as control flow also make it slightly awkward to handle cases where
multiple error messages should be attached in a single parsing process
as one would have to aggregate the individual messages and the raise a
single exception based on those. This is counter intuitive for some
developers as usually exceptions are raise immediately if a
calculation/function cannot be continued. The approach take where the
parser yields a result of one of two types (`ParsingResult[T]` or
`ParsingFailure`) might look a little bit unsual at first glance but
allows the reader to infer the specification of a form field parser by
looking at its type signature. An exception based control flow would
not allow that and neither would it allow a type checker to assist
debugging.

It was also considered to remove the `ParsingResult` type and let
`parse_formfield` return the parsed value directly. This would
complicate cases where `None` is a valid parsing result though and was
therefor rejected.

Currently a `FieldParser` is expected to be a callable. We considered
changing the interface so that there must be a named method a la
`parse_form_field(self, value: str, /) -> ParsingSuccess[T] |
ParsingFailure`. This was rejected as utilizing the `__call__` procol
allows users to pass in functions for parsing as well as instances of
custom classes. The danger of accidentally passing a non-parser
function into `parse_formfields` without the type checker noticing was
considered very very low as such a function would need to yield a
`ParsingSuccess[T] | ParsingFailure` to be accepted.