Change README on how to run flask
The bug in werkzeug seems to be fixed, so we can use `flask run -h localhost` without `python -m`. Moreover, when specifying the server name as "127.0.0.1" instead of "localhost", we can simply run `flask run`. The README has been adapted. You might want to test this with your setup. Moreover, a hint has been added, that it might be necessary to install nix-direnv for the `use flake` line in .envrc to work (this depends on the version of direnv installed).