Refactor QueryPrivateConsumptions
Before this change the QueryPrivateConsumptions use case and its
associated web logic did not conform to our design guidelines. We
addressed this issue by introducing a proper controller
`QueryPrivateConsumptionsController` and updating the call conventions
of the use case to accept a `Request` object, return a `Response`
object and have a named method for calling into the use case instead
of making the use case a callable object.