Include url to transfer request in mail
When requesting a coordination transfer request, a mail gets sent out to the candidate. With this change, the mail contains a link to the  `show_coordination_transfer_request` route, where the candidate can accept the request. This mail to the candiate is currently the only place in the app, where a link to that page is provided.