Do not show name of workers in prd accounts
Before this change the name of the worker that consumed a product was shown in the prd account details of the "selling" company. Now there is shown a string "Anonymous worker" instead. Company names are still shown.

In the future we might show some transaction id to buyer and selling company to facilitate the identification of the buyer.