Remove TransactionGenerator from test code
Before this commit we were using instances of TransactionGenerator in
a few parts of our test code. This transaction generator complicated
the work on other parts of our code base as it exposed operational
internals to the tester that are not part of the business logic but
rather concerns of data representation in the system.

This change removes the TransactionGenerator class from our code and
replaces all instances by direct calls to the database. Those
replacements took place exclusivly in our test code for the database
itself. This should be okay since those tests *should* be concerned
with data representation anyway.