Changed semantics when changing users email address
Before this change a change in a users email address would be
requested by specifying the old email address of that account and the
new email address that should be associated with that account. After
this change we change the email address of a user by specifying their
user id and the new email address in question.

The reason for this change is that our current authentication model is
based on user ids instead of email addresses. Having the old email
address as an authentication token would introduce a syncronization
problem. If somebody would change their email address to an address
that was previously taken by another user they could potentially
authenticated as the previous owner of that email address. This is
true at least as long as we store the email address of a user in the
session cookie. With user ids we will never run into such a problem
because we will never reuse old user ids.

Another way to deal with this problem would have been to check if the
email address that was stored in the session cookie of a user is still
current but that would introduce a lot of overhead and at least one
additional database query for every request.