Introduce ReviewRegisteredConsumptionsUseCase
This use case should be accessible only to an authenticated company and provides itself an overview over the registered consumptions of its products or services ("sales"). This use case is necessary because in the public prd account overview of companies the private consumer details should be anonymized. With this new use case a company can see the details of its customers.

See https://github.com/arbeitszeit/arbeitszeitapp/pull/903#issuecomment-1873423659