Consolidate query_plans views into one single view
Before this change we had 2 separate views/routes for querying all
active plans, one for members and one for companies.  This change
consolidates both of those views into a single one. This will allow
easy integration for accountants in the future as it *should* work out
of the box for that user role.

The business logic rationale is that every authenticated user is part
of the labour time economy, hence they should be able query active
plans.