Implement basic mechanism to update password hashes in DB
Starting with this commit users will automatically update the password
hash stored in our database whenever they log in successfully,
provided that the old hashing algorithm/implementation was considered
out of date by our security provided "werkzeug". This includes
currently only the "sha256" method. All future password hashes will be
generated by using default method provided by "werkzeug".

With the introduction of a new hashing algorithm it became necessary
to extend the length of the `password` field in our database from 100
characters to 300 characters. The length of 300 for this field is
based on a lazy guess and some manual experiments.